requirements.txt
chromadb
sentence-transformers
ollama
tiktoken
einops
python-dotenv
rich
flask
requests